✨ feat: Início da implementação de Ordens de Serviço e Orçamentos
- Adicionados DTOs e serviços iniciais para Ordens de Serviço.
- Implementada estrutura básica para cadastro de Orçamentos.
- Configurada lógica de inclusão e exibição de itens no orçamento.
- Ajustada interface inicial para Ordens de Serviço e Orçamentos.

+++ FastBox.DAL/Models/ItemOrcamento.cs
@@ -15,7 +15,7 @@ public partial class ItemOrcamento
 
     public decimal PrecoUnitario { get; set; }
 
-    public byte Margem { get; set; }
+    public decimal Margem { get; set; }
 
     public virtual Orcamento Orcamento { get; set; } = null!;
 }
